refactor(generate-api-catalog): split main.rs by pipeline stage
The catalog-generation tool's main.rs mixed GitHub discovery, $ref
dereferencing, OpenAPI operation processing, and GraphQL schema
synthesis in one 2407-line file. Split into manifest/github/
dereference/openapi/graphql.rs by pipeline stage; main.rs keeps only
the top-level orchestration.
